+++ husky-emed/husky-emed-ch-cda/src/main/java/org/husky/emed/ch/cda/services/EmedEntryDigestService.java
@@ -42,6 +42,6 @@ public interface EmedEntryDigestService {
      * @param documentationTime     The documentation time.
      * @return The next number in the sequence.
      */
-    long getSequence(final UUID medicationTreatmentId,
-                     final Instant documentationTime);
+    int getSequence(final UUID medicationTreatmentId,
+                    final Instant documentationTime);
 }
